manager/tools/hashtag_analyzer/requirement.txt
selenium>=4.0.0
beautifulsoup4>=4.10.0
requests>=2.25.0
pillow>=8.0.0
webdriver-manager>=3.8.0
opencv-python
moviepy
imageio
imageio-ffmpeg
google-generativeai